requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fastapi==0.110.1
idna==3.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
typing_extensions==4.11.0
uritemplate==4.1.1
